Add files via upload
This project aimed to accurately measure individual load consumption for billing purposes. It involved connecting an energy meter to an Opto-coupler, which counted LED pulses. An 8051 microcontroller read the energy meter via an Opto-isolator and displayed the reading on an LCD. The microcontroller also sent the reading to the user's cell phone through a GSM modem, with the option to auto-save the desired reading through a missed call. The power supply included a step-down transformer, rectifier, and voltage regulator.